+++ admin/src/components/filters-system/exemple/test_volontaire.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-unused-modules */
 import React, { useState } from "react";
 import { translateGrade, youngExportFields } from "snu-lib";